Initial splitting up
Some cleanup

clang-format

Fixes, move GenerateBits/Unrolled to separate header

Fix R compilation

Fix rebase conflict
