+++ cpp/src/arrow/array/data.h
@@ -25,7 +25,6 @@
 
 #include "arrow/buffer.h"
 #include "arrow/result.h"
-#include "arrow/type_fwd.h"
 #include "arrow/util/macros.h"
 #include "arrow/util/visibility.h"
