+++ cpp/src/arrow/compute/api_scalar.h
@@ -20,7 +20,6 @@
 
 #pragma once
 
-#include <memory>
 #include <string>
 #include <utility>
