+++ cpp/src/arrow/compute/kernels/scalar_boolean.cc
@@ -18,8 +18,14 @@
 #include <array>
 
 #include "arrow/compute/kernels/common.h"
+#include "arrow/util/bit_util.h"
+#include "arrow/util/bitmap.h"
+#include "arrow/util/bitmap_ops.h"
 
 namespace arrow {
+
+using internal::Bitmap;
+
 namespace compute {
 
 namespace {
