+++ cpp/src/arrow/compute/kernels/scalar_cast_temporal.cc
@@ -18,11 +18,10 @@
 // Implementation of casting to (or between) temporal types
 
 #include <limits>
-#include <utility>
-#include <vector>
 
 #include "arrow/compute/kernels/common.h"
 #include "arrow/compute/kernels/scalar_cast_internal.h"
+#include "arrow/util/bitmap_reader.h"
 #include "arrow/util/time.h"
 #include "arrow/util/value_parsing.h"
