+++ cpp/src/parquet/platform.h
@@ -20,15 +20,16 @@
 #include <cstdint>
 #include <memory>
 
-#include "arrow/buffer.h"            // IWYU pragma: export
-#include "arrow/io/interfaces.h"     // IWYU pragma: export
-#include "arrow/io/memory.h"         // IWYU pragma: export
-#include "arrow/memory_pool.h"       // IWYU pragma: export
-#include "arrow/status.h"            // IWYU pragma: export
-#include "arrow/util/bit_util.h"     // IWYU pragma: export
-#include "arrow/util/compression.h"  // IWYU pragma: export
-#include "arrow/util/macros.h"       // IWYU pragma: export
-#include "arrow/util/string_view.h"  // IWYU pragma: export
+#include "arrow/buffer.h"              // IWYU pragma: export
+#include "arrow/io/interfaces.h"       // IWYU pragma: export
+#include "arrow/io/memory.h"           // IWYU pragma: export
+#include "arrow/memory_pool.h"         // IWYU pragma: export
+#include "arrow/status.h"              // IWYU pragma: export
+#include "arrow/util/bit_util.h"       // IWYU pragma: export
+#include "arrow/util/bitmap_writer.h"  // IWYU pragma: export
+#include "arrow/util/compression.h"    // IWYU pragma: export
+#include "arrow/util/macros.h"         // IWYU pragma: export
+#include "arrow/util/string_view.h"    // IWYU pragma: export
 
 #if defined(_WIN32) || defined(__CYGWIN__)
