+++ r/src/array_from_vector.cpp
@@ -23,6 +23,7 @@
 #include <arrow/array.h>
 #include <arrow/builder.h>
 #include <arrow/table.h>
+#include <arrow/util/bitmap_writer.h>
 // This says "Private header, not to be exported"
 #include <arrow/visitor_inline.h>
