+++ r/src/array_to_vector.cpp
@@ -20,6 +20,8 @@
 
 #include <arrow/array.h>
 #include <arrow/table.h>
+#include <arrow/util/bitmap_reader.h>
+#include <arrow/util/bitmap_writer.h>
 #include <arrow/util/parallel.h>
 #include <arrow/util/task_group.h>
